+++ backend/SmartSofto.Commerce.Api/ViewModels/CartPriceViewModel.cs
@@ -4,7 +4,7 @@ namespace SmartSofto.Commerce.Api.ViewModels
 {
     public class CartPriceViewModel
     {
-        public int TotalItems { get; set; }
+        public decimal TotalItems { get; set; }
         public decimal Subtotal { get; set; }
         public List<CartPriceItemViewModel> Items { get; set; } = new();
     }
@@ -13,7 +13,7 @@ namespace SmartSofto.Commerce.Api.ViewModels
     {
         public int ProductId { get; set; }
         public string? ProductName { get; set; }
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public decimal UnitPrice { get; set; }
         public decimal LineTotal { get; set; }
     }
